+++ src/bitstream.rs
@@ -7,5 +7,6 @@ pub mod av1;
 pub mod h264;
 pub mod h265;
 mod nal;
+pub mod opus;
 pub mod vp8;
 pub mod vp9;